core: reorganize package
